Support YCQL's JSONB column type yugabyte/yugabyte-db#3953
Summary:
Handle JSONB columns in Spark connector in the same way as we do in the
ycql java driver -- by mapping it to plain String.
Then, the user/developer can use any JSON library to parse and manipulate
the resulting value if needed.

Test Plan: Update TestSparkLocality in yugabyte-db repo

Reviewers: neha

Reviewed By: neha

Subscribers: yql

Differential Revision: https://phabricator.dev.yugabyte.com/D8339